api/requirements.txt
fastapi
google-genai
python-dotenv
uvicorn